[foreman] split foreman plugin
Multiple commands and collecting is executed also
on foreman proxies (Capsules) which ends with error
as for example foreman proxy does not contain
postgresql.

This change should split commands for foreman
and commands for foreman proxy. This means
on foreman/Satellite will be executed both
plugins foreman-proxy and foreman, but on foreman
proxy/Capsule will be collected only foreman-proxy
plugin.

Signed-off-by: Jan Jansky <jjansky@redhat.com>